requirements.txt

fastapi==0.121.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1
python-multipart==0.0.6
